+++ src/components/routes/routeDetails/RouteDetails.module.css
@@ -1,9 +1,3 @@
 .details{
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    text-align: left;
     
 }
