+++ src/components/routes/routeList/RouteList.module.css
@@ -1,6 +1,5 @@
 .routeList{
-    grid-column: 1 / 2;
-    grid-row: 2 / 3;
+    
     display: flex;
     flex-direction: column;
     overflow-y: scroll;
@@ -11,6 +10,7 @@
     scrollbar-color: var(--color-primary) transparent;
     scrollbar-base-color: var(--color-primary);
     scroll-behavior: smooth;
+    
 }
 
 ::-webkit-scrollbar {
